fix(person): fix bug in JLabel removal for Person objects
This commit resolves a bug in removing a JLabel associated with a Person object. If getColony() is null for a Person object, we now search for and remove the JLabel from all four JPanels. This situation doesn't usually occur, but can happen.

feat(movement-strategy): add potentialTargetVec method

In MovementStrategy, I added a method called potentialTargetVec(). This method calculates the next position of the focused object based on the clicked field on the map.

feat(movement-strategy): add processMessage method

In MovementStrategy, I also added a method called processMessage(). This method receives all the delivered messages related to positions and sets the potentialTarget to the last read position (our last click on the map).

feat(cell): add left and right click event listeners

In Cell, I added listeners for the left click (this is to search for a new object on the map to listen to) and for the right click (this is the new target position of the listened object).